requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
isort==6.0.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pre_commit==4.2.0
pydantic==2.11.3
pydantic_core==2.33.1
python-decouple==3.8
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
virtualenv==20.30.0
